Bugfix: Basic properties tried to apply null props
Only apply properties from RabbitMessage to IBasicProperties if there is
any value. Before this change it the BasicPropertiesHandler either
applied the value from the RabbitMessage (if not null) or used the value
present in IBasicProperties. This caused the IBasicProperties class to
register changes to all properties which in turn meant that it tried to
apply all the values when publishing to rabbitMQ. This caused errors
converting null strings to bytes etc.

Now only those properties which are not null are applied to the
IBasicProperties class so only those values which are needed are sent.