Owlery project is a library
Specify in the csproj file that the Owlery project is a library and is
packageable.

When I tried to build using dotnet build on the same level as sln I got
an error message that there was no main function. The OutputType element
fixes that.

When I tried to run dotnet pack on the same level as sln I got an error
that all projects were ignored as none were marked packable. The
IsPackable fixes that.